feat: add delay middleware
This is a middleware that adds a delay to an HTTP request. The
middleware reads the value of the "delay" query parameter
from the URL of the incoming request. If the "delay" parameter
is present, the middleware parses the duration from its value
and waits for the specified duration using the time.Sleep function.
If the duration value is not valid, the middleware returns a
"400 Bad Request" HTTP response with an error message.

Example:
- /api/v1/ping?delay=1s