feat: add status code middleware
A middleware that can be used to manipulate the HTTP response status code.
It inspects the query parameters of the URL in the request. If the
"status_code" query parameter is present and its value is a valid HTTP
status code between 100 and 599, the middleware sets the status code
of the HTTP response to the specified value.